Rename Testshot-mentioning elements and variables
