+++ server/index.ejs
@@ -1,11 +1,11 @@
 <!doctype html>
 <html>
   <head>
-    <title>Testshot</title>
+    <title>Tessereact</title>
   </head>
   <body>
     <script>
-      window.__testshotWSURL = "<%= wsURL %>"
+      window.__tessereactWSURL = "<%= wsURL %>"
       window.__tessereactServerPort = <%= tessereactServerPort %>
     </script>
     <script src="<%= entryPath %>"></script>
