+++ server/index.js
@@ -91,7 +91,7 @@ module.exports = function server (cwd, config, callback) {
             wss.on('connection', (ws) => {
               console.log('Connected to WS')
               ws.on('message', (message) => {
-                console.log('Got message from Testshot runner')
+                console.log('Got message from Tessereact runner')
                 kill()
 
                 if (message === 'OK') {
