+++ src/components/MainView/_lib/prepareStyles/index.js
