+++ src/components/MainView/_lib/scenarios/index.js
