--- src/init/test.js
@@ -1,6 +1,0 @@
-import test from 'ava'
-import init from '.'
-
-test(t => {
-  t.is(typeof init, 'function')
-})
